+++ src/funcs/channelsTargetsUpdate.ts
@@ -56,7 +56,9 @@ export async function channelsTargetsUpdate(
     return parsed;
   }
   const payload = parsed.value;
-  const body = encodeJSON("body", payload.ChannelTarget, { explode: true });
+  const body = encodeJSON("body", payload.ChannelTargetUpdateRequest, {
+    explode: true,
+  });
 
   const pathParams = {
     channel_id: encodeSimple("channel_id", payload.channel_id, {
@@ -98,6 +100,7 @@ export async function channelsTargetsUpdate(
   const requestRes = client._createRequest(context, {
     security: requestSecurity,
     method: "PUT",
+    baseURL: options?.serverURL,
     path: path,
     headers: headers,
     body: body,