+++ src/funcs/directoryUpdate.ts
@@ -92,6 +92,7 @@ export async function directoryUpdate(
   const requestRes = client._createRequest(context, {
     security: requestSecurity,
     method: "PUT",
+    baseURL: options?.serverURL,
     path: path,
     headers: headers,
     body: body,