+++ src/lib/config.ts
@@ -12,13 +12,13 @@ import { Params, pathToFunc } from "./url.js";
  */
 export const ServerList = [
   /**
-   * dev server
+   * API server
    */
-  "http://localhost:8001",
+  "https://api.syllable.ai",
   /**
-   * dev / port-forwarding server
+   * Trial API server
    */
-  "http://localhost:8001",
+  "https://trial.api.syllable.ai",
 ] as const;
 
 export type SDKOptions = {
@@ -61,8 +61,8 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "0.0.2",
-  sdkVersion: "0.1.0-alpha.22",
-  genVersion: "2.477.4",
+  sdkVersion: "0.1.0-alpha.28",
+  genVersion: "2.481.0",
   userAgent:
-    "speakeasy-sdk/typescript 0.1.0-alpha.22 2.477.4 0.0.2 syllable-sdk",
+    "speakeasy-sdk/typescript 0.1.0-alpha.28 2.481.0 0.0.2 syllable-sdk",
 } as const;