+++ src/models/components/channeltarget.ts
@@ -7,16 +7,28 @@ import { remap as remap$ } from "../../lib/primitives.js";
 import { safeParse } from "../../lib/schemas.js";
 import { Result as SafeParseResult } from "../../types/fp.js";
 import { SDKValidationError } from "../errors/sdkvalidationerror.js";
+import {
+  DaoChannelTargetAgent,
+  DaoChannelTargetAgent$inboundSchema,
+  DaoChannelTargetAgent$Outbound,
+  DaoChannelTargetAgent$outboundSchema,
+} from "./daochanneltargetagent.js";
 
+/**
+ * Class for all Channel Target responses
+ */
 export type ChannelTarget = {
   id: number;
-  channelId: number | null;
-  channelName: string | null;
-  agentId: number | null;
-  target: string | null;
-  targetMode: string | null;
-  fallbackTarget: string | null;
-  isTest: boolean | null;
+  channelId: number;
+  channelName: string;
+  agentId: number;
+  agent: DaoChannelTargetAgent;
+  target: string;
+  targetMode: string;
+  fallbackTarget?: string | null | undefined;
+  isTest?: boolean | undefined;
+  updatedAt: Date | null;
+  lastUpdatedBy: string | null;
 };
 
 /** @internal */
@@ -26,13 +38,18 @@ export const ChannelTarget$inboundSchema: z.ZodType<
   unknown
 > = z.object({
   id: z.number().int(),
-  channel_id: z.nullable(z.number().int()),
-  channel_name: z.nullable(z.string()),
-  agent_id: z.nullable(z.number().int()),
-  target: z.nullable(z.string()),
-  target_mode: z.nullable(z.string()),
-  fallback_target: z.nullable(z.string()),
-  is_test: z.nullable(z.boolean()),
+  channel_id: z.number().int(),
+  channel_name: z.string(),
+  agent_id: z.number().int(),
+  agent: DaoChannelTargetAgent$inboundSchema,
+  target: z.string(),
+  target_mode: z.string(),
+  fallback_target: z.nullable(z.string()).optional(),
+  is_test: z.boolean().default(false),
+  updated_at: z.nullable(
+    z.string().datetime({ offset: true }).transform(v => new Date(v)),
+  ),
+  last_updated_by: z.nullable(z.string()),
 }).transform((v) => {
   return remap$(v, {
     "channel_id": "channelId",
@@ -41,19 +58,24 @@ export const ChannelTarget$inboundSchema: z.ZodType<
     "target_mode": "targetMode",
     "fallback_target": "fallbackTarget",
     "is_test": "isTest",
+    "updated_at": "updatedAt",
+    "last_updated_by": "lastUpdatedBy",
   });
 });
 
 /** @internal */
 export type ChannelTarget$Outbound = {
   id: number;
-  channel_id: number | null;
-  channel_name: string | null;
-  agent_id: number | null;
-  target: string | null;
-  target_mode: string | null;
-  fallback_target: string | null;
-  is_test: boolean | null;
+  channel_id: number;
+  channel_name: string;
+  agent_id: number;
+  agent: DaoChannelTargetAgent$Outbound;
+  target: string;
+  target_mode: string;
+  fallback_target?: string | null | undefined;
+  is_test: boolean;
+  updated_at: string | null;
+  last_updated_by: string | null;
 };
 
 /** @internal */
@@ -63,13 +85,16 @@ export const ChannelTarget$outboundSchema: z.ZodType<
   ChannelTarget
 > = z.object({
   id: z.number().int(),
-  channelId: z.nullable(z.number().int()),
-  channelName: z.nullable(z.string()),
-  agentId: z.nullable(z.number().int()),
-  target: z.nullable(z.string()),
-  targetMode: z.nullable(z.string()),
-  fallbackTarget: z.nullable(z.string()),
-  isTest: z.nullable(z.boolean()),
+  channelId: z.number().int(),
+  channelName: z.string(),
+  agentId: z.number().int(),
+  agent: DaoChannelTargetAgent$outboundSchema,
+  target: z.string(),
+  targetMode: z.string(),
+  fallbackTarget: z.nullable(z.string()).optional(),
+  isTest: z.boolean().default(false),
+  updatedAt: z.nullable(z.date().transform(v => v.toISOString())),
+  lastUpdatedBy: z.nullable(z.string()),
 }).transform((v) => {
   return remap$(v, {
     channelId: "channel_id",
@@ -78,6 +103,8 @@ export const ChannelTarget$outboundSchema: z.ZodType<
     targetMode: "target_mode",
     fallbackTarget: "fallback_target",
     isTest: "is_test",
+    updatedAt: "updated_at",
+    lastUpdatedBy: "last_updated_by",
   });
 });